+++ packages/stream_chat/lib/src/core/models/user_group.g.dart
@@ -11,9 +11,7 @@ UserGroup _$UserGroupFromJson(Map<String, dynamic> json) => UserGroup(
   createdBy: json['created_by'] as String?,
   description: json['description'] as String?,
   id: json['id'] as String,
-  members: (json['members'] as List<dynamic>?)
-      ?.map((e) => UserGroupMember.fromJson(e as Map<String, dynamic>))
-      .toList(),
+  members: _membersFromData(json['members'] as List?),
   name: json['name'] as String,
   teamId: json['team_id'] as String?,
   updatedAt: DateTime.parse(json['updated_at'] as String),
@@ -24,7 +22,7 @@ Map<String, dynamic> _$UserGroupToJson(UserGroup instance) => <String, dynamic>{
   'created_by': ?instance.createdBy,
   'description': ?instance.description,
   'id': instance.id,
-  'members': ?instance.members?.map((e) => e.toJson()).toList(),
+  'members': ?_membersToData(instance.members),
   'name': instance.name,
   'team_id': ?instance.teamId,
   'updated_at': instance.updatedAt.toIso8601String(),