+++ packages/stream_chat/test/src/mocks.dart
@@ -13,7 +13,6 @@ import 'package:stream_chat/src/core/api/moderation_api.dart';
 import 'package:stream_chat/src/core/api/polls_api.dart';
 import 'package:stream_chat/src/core/api/push_preferences_api.dart';
 import 'package:stream_chat/src/core/api/user_api.dart';
-import 'package:stream_chat/src/core/api/user_groups_api.dart';
 import 'package:stream_chat/src/core/http/stream_http_client.dart';
 import 'package:stream_chat/src/core/models/channel_config.dart';
 import 'package:stream_chat/src/core/util/event_controller.dart';
@@ -56,8 +55,6 @@ class MockChannelApi extends Mock implements ChannelApi {}
 class MockPushPreferencesApi extends Mock implements PushPreferencesApi {}
 
 class MockModerationApi extends Mock implements ModerationApi {}
-
-class MockUserGroupsApi extends Mock implements UserGroupsApi {}
 
 class MockDefaultApi extends Mock implements DefaultApi {}